refactor: split functions and types to separate files
