+++ src/types.js
@@ -1,0 +1,48 @@
+/**
+ * @typedef Options
+ *   Configuration.
+ * @property {string} [bibliography]
+ *   Name of bibtex or CSL-JSON file
+ * @property {string} [path]
+ *   Optional path to file (node). Will be joined with `options.bibliography` and used in place of cwd of file if provided.
+ * @property {'apa'|'vancouver'|'harvard1'|'chicago'|'mla'|string} [csl]
+ *   One of 'apa', 'vancouver', 'harvard1', 'chicago', 'mla'. A local file path or URL to a valid CSL file is also accepted.
+ * @property {string} [lang]
+ *   Locale to use in formatting citations. Defaults to en-US. A local file path or URL to a valid locale file is also accepted.
+ * @property {boolean} [suppressBibliography]
+ *   By default, biliography is inserted after the entire markdown file.
+ *   If the file contains `[^Ref]`, the biliography will be inserted there instead.
+ * @property {boolean} [linkCitations]
+ *   If true, citations will be hyperlinked to the corresponding bibliography entries (for author-date and numeric styles only).
+ *   Defaults to false.
+ * @property {string[]} [noCite]
+ *   Citation IDs (@item1) to include in the bibliography even if they are not cited in the document
+ * @property {string[]} [inlineClass]
+ *   Class(es) to add to the inline citation.
+ * @property {string[]} [inlineBibClass]
+ *   Class(es) to add to the inline bibliography. Leave empty for no inline bibliography.
+ */
+
+/**
+ * @typedef CiteItem
+ *   Cite item to be passed into citeproc-js
+ * @property {string} [id]
+ *   The id field is required
+ * @property {string} [locator]
+ *   A string identifying a page number or other pinpoint location or range within the resource;
+ * @property {string} [label]
+ *   Path to file
+ * @property {string} [prefix]
+ *   A string to print before this cite item
+ * @property {string} [suffix]
+ *   A string to print after this cite item
+ * @property {boolean} [suppress-author]
+ *   If true, author names will not be included in the citation output for this cite
+ * @property {boolean} [author-only]
+ *   If true, only the author name will be included in the citation output for this cite
+ *
+ * @typedef {"note" | "in-text"} Mode
+ * @typedef {'author-date' | 'author' | 'numeric' | 'note' | 'label'} CitationFormat
+ */
+
+export const Types = {}
